[REFACTOR] CQRS패턴 적용 #266
